requirements.txt
asgiref==3.7.2
Django==4.2.8
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.8.0
